+++ sdk/core/azure-core/src/io/body_stream.cpp
@@ -17,7 +17,7 @@
 #endif
 
 #include "azure/core/context.hpp"
-#include "azure/core/http/body_stream.hpp"
+#include "azure/core/io/body_stream.hpp"
 
 #include <algorithm>
 #include <cstdint>
@@ -28,7 +28,7 @@
 #include <vector>
 
 using Azure::Core::Context;
-using namespace Azure::Core::Http;
+using namespace Azure::IO;
 
 // Keep reading until buffer is all fill out of the end of stream content is reached
 int64_t BodyStream::ReadToCount(
@@ -133,13 +133,3 @@ int64_t FileBodyStream::OnRead(Azure::Core::Context const& context, uint8_t* buf
   return numberOfBytesRead;
 }
 #endif
-
-int64_t LimitBodyStream::OnRead(Context const& context, uint8_t* buffer, int64_t count)
-{
-  (void)context;
-  // Read up to count or whatever length is remaining; whichever is less
-  uint64_t bytesRead
-      = m_inner->Read(context, buffer, std::min(count, this->m_length - this->m_bytesRead));
-  this->m_bytesRead += bytesRead;
-  return bytesRead;
-}