+++ sdk/storage/azure-storage-blobs/test/ut/block_blob_client_test.cpp
@@ -50,8 +50,7 @@ namespace Azure { namespace Storage { namespace Test {
     m_blobUploadOptions.HttpHeaders.ContentEncoding = "identity";
     m_blobUploadOptions.HttpHeaders.ContentHash.Value.clear();
     m_blobUploadOptions.Tier = Azure::Storage::Blobs::Models::AccessTier::Hot;
-    auto blobContent
-        = Azure::Core::Http::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
+    auto blobContent = Azure::IO::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
     m_blockBlobClient->Upload(&blobContent, m_blobUploadOptions);
     m_blobUploadOptions.HttpHeaders.ContentHash
         = m_blockBlobClient->GetProperties()->HttpHeaders.ContentHash;
@@ -63,8 +62,7 @@ namespace Azure { namespace Storage { namespace Test {
   {
     auto blockBlobClient = Azure::Storage::Blobs::BlockBlobClient::CreateFromConnectionString(
         StandardStorageConnectionString(), m_containerName, RandomString());
-    auto blobContent
-        = Azure::Core::Http::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
+    auto blobContent = Azure::IO::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
     auto blobContentInfo = blockBlobClient.Upload(&blobContent, m_blobUploadOptions);
     EXPECT_FALSE(blobContentInfo->RequestId.empty());
     EXPECT_TRUE(blobContentInfo->ETag.HasValue());
@@ -179,8 +177,7 @@ namespace Azure { namespace Storage { namespace Test {
     std::vector<uint8_t> emptyContent;
     auto blockBlobClient = Azure::Storage::Blobs::BlockBlobClient::CreateFromConnectionString(
         StandardStorageConnectionString(), m_containerName, RandomString());
-    auto blobContent
-        = Azure::Core::Http::MemoryBodyStream(emptyContent.data(), emptyContent.size());
+    auto blobContent = Azure::IO::MemoryBodyStream(emptyContent.data(), emptyContent.size());
     blockBlobClient.Upload(&blobContent);
     blockBlobClient.SetHttpHeaders(m_blobUploadOptions.HttpHeaders);
     blockBlobClient.SetMetadata(m_blobUploadOptions.Metadata);
@@ -208,20 +205,17 @@ namespace Azure { namespace Storage { namespace Test {
   {
     auto blobClient = m_blobContainerClient->GetBlobClient(RandomString());
     auto res = blobClient.StartCopyFromUri(m_blockBlobClient->GetUrl());
-    EXPECT_FALSE(res->RequestId.empty());
-    EXPECT_FALSE(res.GetRawResponse().GetHeaders().at(Details::HttpHeaderRequestId).empty());
-    EXPECT_FALSE(res.GetRawResponse().GetHeaders().at(Details::HttpHeaderDate).empty());
-    EXPECT_FALSE(res.GetRawResponse().GetHeaders().at(Details::HttpHeaderXMsVersion).empty());
-    EXPECT_TRUE(res->ETag.HasValue());
-    EXPECT_TRUE(IsValidTime(res->LastModified));
-    EXPECT_FALSE(res->CopyId.empty());
-    EXPECT_TRUE(res->VersionId.HasValue());
-    EXPECT_FALSE(res->VersionId.GetValue().empty());
+    EXPECT_FALSE(res.RequestId.empty());
+    EXPECT_TRUE(res.ETag.HasValue());
+    EXPECT_TRUE(IsValidTime(res.LastModified));
+    EXPECT_FALSE(res.CopyId.empty());
+    EXPECT_TRUE(res.VersionId.HasValue());
+    EXPECT_FALSE(res.VersionId.GetValue().empty());
     EXPECT_TRUE(
-        res->CopyStatus == Azure::Storage::Blobs::Models::CopyStatus::Pending
-        || res->CopyStatus == Azure::Storage::Blobs::Models::CopyStatus::Success);
+        res.CopyStatus == Azure::Storage::Blobs::Models::CopyStatus::Pending
+        || res.CopyStatus == Azure::Storage::Blobs::Models::CopyStatus::Success);
     auto properties = *blobClient.GetProperties();
-    EXPECT_EQ(properties.CopyId.GetValue(), res->CopyId);
+    EXPECT_EQ(properties.CopyId.GetValue(), res.CopyId);
     EXPECT_FALSE(properties.CopySource.GetValue().empty());
     EXPECT_TRUE(
         properties.CopyStatus.GetValue() == Azure::Storage::Blobs::Models::CopyStatus::Pending
@@ -236,7 +230,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_FALSE(properties.IncrementalCopyDestinationSnapshot.HasValue());
 
     auto downloadResult = blobClient.Download();
-    EXPECT_EQ(downloadResult->Details.CopyId.GetValue(), res->CopyId);
+    EXPECT_EQ(downloadResult->Details.CopyId.GetValue(), res.CopyId);
     EXPECT_FALSE(downloadResult->Details.CopySource.GetValue().empty());
     EXPECT_TRUE(
         downloadResult->Details.CopyStatus.GetValue()
@@ -271,7 +265,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_EQ(ReadBodyStream(versionClient.Download()->BodyStream), m_blobContent);
     EXPECT_EQ(versionClient.GetProperties()->Metadata, m_blobUploadOptions.Metadata);
     EXPECT_TRUE(versionClient.GetProperties()->IsServerEncrypted);
-    auto emptyContent = Azure::Core::Http::MemoryBodyStream(nullptr, 0);
+    auto emptyContent = Azure::IO::MemoryBodyStream(nullptr, 0);
     EXPECT_THROW(snapshotClient.Upload(&emptyContent), StorageException);
     EXPECT_THROW(snapshotClient.SetMetadata({}), StorageException);
     EXPECT_NO_THROW(snapshotClient.SetAccessTier(Azure::Storage::Blobs::Models::AccessTier::Cool));
@@ -367,8 +361,7 @@ namespace Azure { namespace Storage { namespace Test {
   {
     auto blockBlobClient = Azure::Storage::Blobs::BlockBlobClient::CreateFromConnectionString(
         StandardStorageConnectionString(), m_containerName, RandomString());
-    auto blobContent
-        = Azure::Core::Http::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
+    auto blobContent = Azure::IO::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
     blockBlobClient.Upload(&blobContent);
     blockBlobClient.SetMetadata(m_blobUploadOptions.Metadata);
     blockBlobClient.SetAccessTier(Azure::Storage::Blobs::Models::AccessTier::Cool);
@@ -399,8 +392,7 @@ namespace Azure { namespace Storage { namespace Test {
     std::vector<uint8_t> block1Content;
     block1Content.resize(100);
     RandomBuffer(reinterpret_cast<char*>(&block1Content[0]), block1Content.size());
-    auto blockContent
-        = Azure::Core::Http::MemoryBodyStream(block1Content.data(), block1Content.size());
+    auto blockContent = Azure::IO::MemoryBodyStream(block1Content.data(), block1Content.size());
     blockBlobClient.StageBlock(blockId1, &blockContent);
     Azure::Storage::Blobs::CommitBlockListOptions options;
     options.HttpHeaders = m_blobUploadOptions.HttpHeaders;
@@ -722,8 +714,7 @@ namespace Azure { namespace Storage { namespace Test {
     std::vector<uint8_t> emptyContent;
     auto blockBlobClient = Azure::Storage::Blobs::BlockBlobClient::CreateFromConnectionString(
         StandardStorageConnectionString(), m_containerName, RandomString());
-    auto blobContent
-        = Azure::Core::Http::MemoryBodyStream(emptyContent.data(), emptyContent.size());
+    auto blobContent = Azure::IO::MemoryBodyStream(emptyContent.data(), emptyContent.size());
     blockBlobClient.Upload(&blobContent);
     blockBlobClient.SetHttpHeaders(m_blobUploadOptions.HttpHeaders);
     blockBlobClient.SetMetadata(m_blobUploadOptions.Metadata);